Set seed type to u64
There are more than 4 billion possible worlds out there!

+++ mehlon-server/config.rs
@@ -4,7 +4,7 @@ use super::StrErr;
 
 #[derive(Deserialize, Clone)]
 pub struct Config {
-	pub mapgen_seed :u32,
+	pub mapgen_seed :u64,
 	pub mapgen_radius_xy :isize,
 	pub mapgen_radius_z :isize,
